+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-08T07:01:32.927Z")
+    created: Optional[datetime] = parse_datetime("2025-10-08T14:08:06.833Z")
     r"""The date and time the resource was created"""