+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -303,7 +303,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K71A095EP49Q7Y3HNTAXCRH2"
+    id: Optional[str] = "01K722DBMJXA45V6CSGF550MZT"
     r"""The id of the resource"""