+++ packages/orq-rc/RELEASES.md
@@ -3209,3 +3209,13 @@ Based on:
 - [python v3.14.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.16] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.16 - packages/orq-rc
+
+## 2025-10-08 14:09:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.1 (2.723.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.17 - packages/orq-rc